reverted some changes so game runs
the changes 2 commits ago where Colors became a module caused errors in Output. Output being a module also can't store instance variables, which was used in order for the Code to be printed at the end. It was terrible implementation in the first place, but now it caused problems. Also forgot to require GameStatus into User at some point. Overall this is all a huge mess, but i need to implement feedback first b4 refactoring again so i can merge this branch